Release 6 – add delact

+++ vmafia.py
@@ -160,7 +160,6 @@ def triggers(msg):
                 if bot.get_chat_member(GROUP_ID, uid).status == 'member':
                     msg_delete = bot.send_message(msg.chat.id, text.addact)
                     cur.execute("INSERT INTO active (uids) VALUES (%s)", [uid])
-                    cur.execute("INSERT INTO all_uids (list) VALUES (%s)", [uid])
                     uids.append(uid)
                     time.sleep(3)
                     bot.delete_message(msg.chat.id, msg.message_id)
@@ -172,36 +171,6 @@ def triggers(msg):
                     bot.delete_message(msg.chat.id, msg_delete.message_id)
             else:
                 msg_delete = bot.send_message(msg.chat.id, text.noaddact)
-                time.sleep(3)
-                bot.delete_message(msg.chat.id, msg.message_id)
-                bot.delete_message(msg.chat.id, msg_delete.message_id)
-        else:
-            msg_delete = bot.send_message(msg.chat.id, text=text.onlyadm.format(msg.from_user.id), parse_mode="HTML")
-            time.sleep(5)
-            bot.delete_message(msg.chat.id, msg.message_id)
-            bot.delete_message(msg.chat.id, msg_delete.message_id)
-
-@bot.message_handler(commands=['del_forever'])
-def triggers(msg):
-    if not msg.chat.id == GROUP_ID:
-        bot.send_message(msg.chat.id, text.notmafia)
-    else:
-        admins = [admin.user.id for admin in bot.get_chat_administrators(GROUP_ID)]
-        if msg.from_user.id in admins:
-            uid = msg.reply_to_message.from_user.id
-            print(uid)
-            if uid in all_uids:
-                msg_delete = bot.send_message(msg.chat.id, text.delact)
-                cur.execute('DELETE FROM all_uids WHERE list = %s', [uid])
-                if uid in uids:
-                    cur.execute('DELETE FROM active WHERE uids = %s', [uid])
-                    uids.remove(uid)
-                conn.commit()
-                time.sleep(3)
-                bot.delete_message(msg.chat.id, msg.message_id)
-                bot.delete_message(msg.chat.id, msg_delete.message_id)
-            else:
-                msg_delete = bot.send_message(msg.chat.id, text.delact)
                 time.sleep(3)
                 bot.delete_message(msg.chat.id, msg.message_id)
                 bot.delete_message(msg.chat.id, msg_delete.message_id)
@@ -224,6 +193,35 @@ def triggers(msg):
                 cur.execute('DELETE FROM active WHERE uids = %s', [uid])
                 conn.commit()
                 uids.remove(uid)
+                time.sleep(3)
+                bot.delete_message(msg.chat.id, msg.message_id)
+                bot.delete_message(msg.chat.id, msg_delete.message_id)
+            else:
+                msg_delete = bot.send_message(msg.chat.id, text.delact)
+                time.sleep(3)
+                bot.delete_message(msg.chat.id, msg.message_id)
+                bot.delete_message(msg.chat.id, msg_delete.message_id)
+        else:
+            msg_delete = bot.send_message(msg.chat.id, text=text.onlyadm.format(msg.from_user.id), parse_mode="HTML")
+            time.sleep(5)
+            bot.delete_message(msg.chat.id, msg.message_id)
+            bot.delete_message(msg.chat.id, msg_delete.message_id)
+
+@bot.message_handler(commands=['del_forever'])
+def triggers(msg):
+    if not msg.chat.id == GROUP_ID:
+        bot.send_message(msg.chat.id, text.notmafia)
+    else:
+        admins = [admin.user.id for admin in bot.get_chat_administrators(GROUP_ID)]
+        if msg.from_user.id in admins:
+            uid = msg.reply_to_message.from_user.id
+            if uid in all_uids:
+                msg_delete = bot.send_message(msg.chat.id, text.delact)
+                cur.execute('DELETE FROM all_uids WHERE list = %s', [uid])
+                if uid in uids:
+                    cur.execute('DELETE FROM active WHERE uids = %s', [uid])
+                    uids.remove(uid)
+                conn.commit()
                 time.sleep(3)
                 bot.delete_message(msg.chat.id, msg.message_id)
                 bot.delete_message(msg.chat.id, msg_delete.message_id)
